Moves fetchAppTokenMiddleware
